requirements.txt
alembic==1.16.5
aniso8601==10.0.1
attrs==25.4.0
Authlib==1.6.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==7.1.2
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
email-validator==1.1.3
filelock==3.19.1
Flask==1.1.4
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fsspec==2025.10.0
greenlet==3.2.4
gunicorn==20.1.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==2.0.1
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pycparser==2.23
PyMySQL==1.1.2
python-decouple==3.4
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rpds-py==0.27.1
safetensors==0.7.0
six==1.17.0
SQLAlchemy==1.4.23
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==1.0.1
WTForms==2.3.3
